requirements.txt
numpy
scipy
matplotlib
opencv-python==3.3.0.0
pims_nd2
python-bioformats
vtk
envisage
wxPython
mayavi